+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Add GIMP for photo editing.
 - Add mpv for HDR video playback.
 - Add ckb-next for supporting RGB on Corsair keyboards and mice.
+- Change winesapOS testing repository to be a rolling updates repository.
 
 ## 4.3.0 - The Wayland Update